+++ next.config.js
@@ -1,6 +1,12 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-}
+  compiler: {
+    styledComponents: true,
+  },
+  eslint: {
+    dirs: ["."],
+  },
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;