+++ three/shaders/vertexShaders/uvPosition.js
@@ -1,4 +1,4 @@
-export default `
+const shader = `
 attribute vec2 textureUV;
 varying vec3 pos;
 
@@ -7,3 +7,5 @@ void main() {
   gl_Position = vec4(textureUV*2.0-1.0,-1.0, 1.0);
 }
 `;
+
+export default shader;